requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
ollama>=0.1.0
tk>=0.1.0
numpy>=1.24.0
llava-torch>=0.7.0